# docs/api_doc.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Library Management Service API",
    "description": "Library Management Service API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Get authentication token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "login/retrieve token successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginPresenter"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/docgia": {
      "get": {
        "tags": [
          "Doc gia"
        ],
        "description": "Get danh sach doc gia",
        "responses": {
          "200": {
            "description": "Get danh sach doc gia successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocGiaPresenter"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "description": "Create new doc gia",
        "tags": [
          "Doc gia"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocGiaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create doc gia successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocGiaPresenter"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/docgia/{maDocGia}": {
      "get": {
        "tags": [
          "Doc gia"
        ],
        "description": "Get doc gia by id",
        "parameters": [
          {
            "in": "path",
            "name": "maDocGia",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get doc gia successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocGiaPresenter"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "description": "Update doc gia successful",
        "tags": [
          "Doc gia"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "maDocGia",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocGiaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update doc gia successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocGiaPresenter"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "Doc gia"
        ],
        "description": "Delete doc gia by id",
        "parameters": [
          {
            "in": "path",
            "name": "maDocGia",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete doc gia successful"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/dausach": {
      "get": {
        "summary": "Get danh sach dau sach",
        "tags": [
          "Dau sach"
        ],
        "responses": {
          "200": {
            "description": "Get danh sach dau sach successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DauSachPresenter"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Create dau sach",
        "tags": [
          "Dau sach"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DauSachDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create dau sach successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DauSachPresenter"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/dausach/{maDauSach}": {
      "get": {
        "summary": "Get dau sach by ma dau sach",
        "tags": [
          "Dau sach"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "maDauSach",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get dau sach successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DauSachPresenter"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "put": {
        "summary": "Update dau sach",
        "tags": [
          "Dau sach"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "maDauSach",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DauSachDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Create dau sach successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DauSachPresenter"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete dau sach by ma dau sach",
        "tags": [
          "Dau sach"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "maDauSach",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete dau sach successful"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test@email.com"
          },
          "password": {
            "type": "string",
            "example": 12345678
          }
        }
      },
      "LoginPresenter": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Jwt token format"
          }
        }
      },
...
      "DocGiaDto": {
        "type": "object",
        "properties": {
          "hoTen": {
            "type": "string"
          },
          "maLoaiDocGia": {
            "type": "string"
          },
          "ngaySinh": {
            "type": "string",
            "format": "date-time"
          },
          "diaChi": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "ngayLapThe": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DocGiaPresenter": {
        "type": "object",
        "properties": {
          "maDocGia": {
            "type": "string"
          },
          "hoTen": {
            "type": "string"
          },
          "loaiDocGia": {
            "type": "string"
          },
          "ngaySinh": {
            "type": "string",
            "format": "date-time"
          },
          "diaChi": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "ngayLapThe": {
            "type": "string",
            "format": "date-time"
          },
          "ngayHetHan": {
            "type": "string",
            "format": "date-time,"
          },
          "tongNo": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "TheLoaiPresenter": {
        "type": "object",
        "properties": {
          "maTheLoai": {
            "type": "string"
          },
          "tenTheLoai": {
            "type": "string"
          }
        }
      },
...
      "TacGiaPresenter": {
        "type": "object",
        "properties": {
          "maTacGia": {
            "type": "string"
          },
          "tenTacGia": {
            "type": "string"
          }
        }
      },
      "DauSachDto": {
        "type": "object",
        "properties": {
          "tenDauSach": {
            "type": "string"
          },
          "theLoai": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tacGia": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "DauSachPresenter": {
        "type": "object",
        "properties": {
          "maDauSach": {
            "type": "string"
          },
          "tenDauSach": {
            "type": "string"
          },
          "theLoai": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TheLoaiPresenter"
            }
          },
          "tacGia": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TacGiaPresenter"
            }
          }
        }
      },
...
    },
    "responses": {
      "BadRequestError": {
        "description": "provide bad request"
      },
      "ForbiddenError": {
        "description": "do not have permission to view this resource"
      },
      "NotFoundError": {
        "description": "resource not found"
      },
      "UnauthorizedError": {
        "description": "token is missing"
      },
      "ConflictError": {
        "description": "conflict with state of server, maybe due to database constraint reason,.."
      },
      "InternalServerError": {
        "description": "internal server error"
      }
    }
  }
}